+++ src/base/b-virtual-scroll/b-virtual-scroll.ts
@@ -6,6 +6,11 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
+/**
+ * [[include:base/b-virtual-scroll/README.md]]
+ * @packageDocumentation
+ */
+
 //#if demo
 import 'models/demo/pagination';
 //#endif
@@ -18,9 +23,11 @@ import iData, {
 
 	component,
 	prop,
-	field,
 	system,
+	field,
+	watch,
 	wait,
+	hook,
 	p,
 
 	RequestParams,
@@ -47,6 +54,8 @@ import {
 	RemoteData,
 	LocalState,
 	RequestQueryFn,
+	DataState,
+	MergeDataStateParams,
 	UnsafeBVirtualScroll
 
 } from 'base/b-virtual-scroll/interface';
@@ -91,24 +100,28 @@ export default class bVirtualScroll extends iData implements iItems {
 	/**
 	 * Maximum number of elements to cache
 	 */
+	// eslint-disable-next-line @typescript-eslint/unbound-method
 	@prop({type: Number, watch: 'syncPropsWatcher', validator: Number.isNatural})
 	readonly cacheSize: number = 400;
 
 	/**
 	 * Number of elements till the page bottom that should initialize a new render iteration
 	 */
+	// eslint-disable-next-line @typescript-eslint/unbound-method
 	@prop({type: Number, validator: Number.isNatural})
 	readonly renderGap: number = 10;
 
 	/**
 	 * Number of elements per one render chunk
 	 */
+	// eslint-disable-next-line @typescript-eslint/unbound-method
 	@prop({type: Number, validator: Number.isNatural})
 	readonly chunkSize: number = 10;
 
 	/**
 	 * Number of tombstones to render
 	 */
+	// eslint-disable-next-line @typescript-eslint/unbound-method
 	@prop({type: Number, required: false, validator: Number.isNatural})
 	readonly tombstonesSize?: number;
 
@@ -150,7 +163,7 @@ export default class bVirtualScroll extends iData implements iItems {
 	/**
 	 * When this function returns true the component will be able to request additional data
 	 */
-	@prop({type: Function, default: (v) => v.itemsTillBottom <= 10 && !v.isLastEmpty})
+	@prop({type: Function, default: (v: DataState) => v.itemsTillBottom <= 10 && !v.isLastEmpty})
 	readonly shouldMakeRequest!: RequestFn;
 
 	/**
@@ -199,15 +212,16 @@ export default class bVirtualScroll extends iData implements iItems {
 	protected get requestParams(): RequestParams {
 		return {
 			get: {
-				...this.requestQuery?.(getRequestParams())?.get,
-				...(<Dictionary<Dictionary>>this.request)?.get
+				...this.requestQuery?.(this.getDataStateSnapshot())?.get,
+				...Object.isDictionary(this.request?.get) ? this.request?.get : undefined
 			}
 		};
 	}
 
 	/** @override */
+	// eslint-disable-next-line @typescript-eslint/no-unused-vars-experimental
 	protected set requestParams(value: RequestParams) {
-		return;
+		// Loopback
 	}
 
 	/**
@@ -238,10 +252,17 @@ export default class bVirtualScroll extends iData implements iItems {
 		done?: HTMLElement;
 	};
 
-	/** @override */
+	/**
+	 * @override
+	 * @emits chunkLoading(page: number)
+	 */
 	initLoad(data?: unknown, params: InitLoadOptions = {}): CanPromise<void> {
 		if (!this.lfc.isBeforeCreate()) {
-			this.reInit().catch(stderr);
+			this.reInit();
+		}
+
+		if (this.isActivated) {
+			this.emit('chunkLoading', 0);
 		}
 
 		return super.initLoad(data, params);
@@ -251,7 +272,7 @@ export default class bVirtualScroll extends iData implements iItems {
 	 * Reloads the last request (if there is no `db` or `options` the method calls reload)
 	 */
 	reloadLast(): void {
-		if (!this.db || !this.options.length) {
+		if (!this.db || this.chunkRequest.data.length === 0) {
 			this.reload().catch(stderr);
 
 		} else {
@@ -262,13 +283,59 @@ export default class bVirtualScroll extends iData implements iItems {
 	/**
 	 * Re-initializes component
 	 */
-	async reInit(): Promise<void> {
+	reInit(): void {
 		this.componentRender.reInit();
 		this.chunkRequest.reset();
 		this.chunkRender.reInit();
 	}
 
-	/** @override  */
+	/**
+	 * Returns an object with the current data state of the component
+	 *
+	 * @typeParam ITEM - data item to render
+	 * @typeParam RAW - raw provider data
+	 */
+	getCurrentDataState<
+		ITEM extends unknown = unknown,
+		RAW extends unknown = unknown
+	>(): DataState<ITEM, RAW> {
+		let overrideParams: MergeDataStateParams = {};
+
+		if (this.componentStatus !== 'ready' || !Object.isTruly(this.dataProvider)) {
+			overrideParams = {
+				currentPage: 0,
+				...overrideParams
+			};
+		}
+
+		return this.getDataStateSnapshot(overrideParams, this.chunkRequest, this.chunkRender);
+	}
+
+	/**
+	 * Takes a snapshot of the current data state and returns it
+	 *
+	 * @param [overrideParams]
+	 * @param [chunkRequest]
+	 * @param [chunkRender]
+	 *
+	 * @typeParam ITEM - data item to render
+	 * @typeParam RAW - raw provider data
+	 */
+	protected getDataStateSnapshot<
+		ITEM extends unknown = unknown,
+		RAW extends unknown = unknown
+	>(
+		overrideParams?: MergeDataStateParams,
+		chunkRequest?: ChunkRequest,
+		chunkRender?: ChunkRender
+	): DataState<ITEM, RAW> {
+		return getRequestParams(chunkRequest, chunkRender, overrideParams);
+	}
+
+	/**
+	 * @override
+	 * @emits chunkLoaded(lastLoadedChunk: LastLoadedChunk)
+	 */
 	protected initRemoteData(): void {
 		if (!this.db) {
 			return;
@@ -277,24 +344,59 @@ export default class bVirtualScroll extends iData implements iItems {
 		this.localState = 'init';
 
 		const
-			val = this.convertDBToComponent<RemoteData>(this.db);
+			{data, total} = this.db;
 
-		if (this.field.get('data.length', val)) {
-			const
-				params = getRequestParams(undefined, undefined, {lastLoadedData: val.data});
+		if (data && data.length > 0) {
+			const lastLoadedChunk = {
+				normalized: data,
+				raw: this.chunkRequest.lastLoadedChunk.raw
+			};
 
+			const params = this.getDataStateSnapshot({
+				lastLoadedData: data,
+				lastLoadedChunk
+			});
+
+			this.chunkRequest.lastLoadedChunk = lastLoadedChunk;
 			this.chunkRequest.shouldStopRequest(params);
-			this.options = val.data;
-			this.total = Object.isNumber(val.total) ? val.total : undefined;
+			this.chunkRequest.data = data;
+			this.total = total;
 
 		} else {
+			this.chunkRequest.isLastEmpty = true;
+
 			const
-				params = getRequestParams(undefined, undefined, {isLastEmpty: true});
+				params = this.getDataStateSnapshot({isLastEmpty: true});
 
 			this.chunkRequest.shouldStopRequest(params);
-			this.options = [];
 		}
 
+		this.emit('chunkLoaded', this.chunkRequest.lastLoadedChunk);
+		this.chunkRequest.init().catch(stderr);
+	}
+
+	/** @override */
+	protected convertDataToDB<O>(data: unknown): O | this['DB'] {
+		this.chunkRequest.lastLoadedChunk.raw = data;
+		return super.convertDataToDB(data);
+	}
+
+	/**
+	 * Initializes rendering on the items passed to the component
+	 */
+	@hook('mounted')
+	@watch('options')
+	@wait('ready', {defer: true, label: $$.initOptions})
+	protected initItems(): CanPromise<void> {
+		if (this.dataProvider !== undefined) {
+			return;
+		}
+
+		if (this.localState === 'ready') {
+			this.reInit();
+		}
+
+		this.chunkRequest.lastLoadedChunk.normalized = Object.isArray(this.options) ? [...this.options] : [];
 		this.chunkRequest.init().catch(stderr);
 	}
 
@@ -304,11 +406,24 @@ export default class bVirtualScroll extends iData implements iItems {
 	}
 
 	/**
-	 * Synchronization for the component props
+	 * Synchronization of the component props
 	 */
 	@wait('ready', {defer: true, label: $$.syncPropsWatcher})
-	protected async syncPropsWatcher(): Promise<void> {
+	protected syncPropsWatcher(): CanPromise<void> {
 		return this.reInit();
+	}
+
+	/** @override */
+	protected syncDataProviderWatcher(): void {
+		const
+			provider = this.dataProvider;
+
+		if (provider === undefined) {
+			this.reInit();
+
+		} else {
+			super.syncDataProviderWatcher();
+		}
 	}
 
 	/** @override */