+++ src/super/i-block/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.40 (2020-07-23)
+
+#### :nail_care: Polish
+
+* Added waitRef to the UnsafeIBlock declaration
+
 ## v3.0.0-rc.37 (2020-07-20)
 
 #### :boom: Breaking Change